data-cleansing/solutions/requirements.txt
pandas==1.5.3
PyMySQL==1.0.2
python-dotenv==0.15.0
